// demos/showcase/supply-chain/types.ts
/****************************************************************************
 ** @license
 ** This demo file is part of yFiles for HTML.
 ** Copyright (c) 2026 by yWorks GmbH, Vor dem Kreuzberg 28,
 ** 72070 Tuebingen, Germany. All rights reserved.
 **
 ** yFiles demo files exhibit yFiles for HTML functionalities. Any redistribution
 ** of demo files in source code or binary form, with or without
 ** modification, is not permitted.
 **
 ** Owners of a valid software license for a yFiles for HTML version that this
 ** demo is shipped with are allowed to use the demo source code as basis
 ** for their own yFiles for HTML powered applications. Use of such programs is
 ** governed by the rights and conditions as set out in the yFiles for HTML
 ** license agreement.
 **
 ** THIS SOFTWARE IS PROVIDED ''AS IS'' AND ANY EXPRESS OR IMPLIED
 ** WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
 ** MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN
 ** NO EVENT SHALL yWorks BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
 ** SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
 ** TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
 ** PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
 ** LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
 ** NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 ** SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 **
 ***************************************************************************/
import type { TypedHtmlVisual } from '@yfiles/yfiles'

export type NodeData = {
  id: string
  headline: string
  location?: string
  icon?: string
  groupId?: string
  layer?: number
  properties: { id: string; name: string; stock: number; endOfChain?: boolean }[]
}

export type EdgeData = {
  id: string
  source: { nodeId: string; propertyId: string }
  target: { nodeId: string; propertyId: string }
}

export type GroupData = { id: string; headline: string }

export type GraphData = { nodes: NodeData[]; edges: EdgeData[]; groups: GroupData[] }

export type NodeTagProperty = {
  elementId: string
  toHighlight: boolean
  toFlash: boolean
  toAlert: boolean
  toFlashAlert: boolean
  stock: number
}

export type HtmlEditableNodeStyleVisual = TypedHtmlVisual<HTMLDivElement>